rename password input
