+++ packages/password-input/src/index.ts
@@ -1,5 +1,5 @@
-export { PasswordInput } from './PasswordInput';
+export { PasswordInput } from './Password-Input';
 
 // types
 
-export type { PasswordInputProps } from './PasswordInput';
+export type { PasswordInputProps } from './Password-Input';
